% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{Frequency table}
\usage{
freqTable(df, var1, var2)
}
\arguments{
\item{df}{a data.frame containing var1 and var2}

\item{var1}{variable to count observations by}

\item{var2}{additional variable to count observations by (optional)}
}
\value{
A tibble with the number of observations per distinct value of var 1 and optionally var2
}
\description{
Creates 1-dimensional or 2-dimensional frequency tables
}
\examples{
data(mpg)
freqTable(mpg, "trans")
freqTable(mpg, "trans", "year")
}
