% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icolors.R
\name{icolors.lookup}
\alias{icolors.lookup}
\title{icolors.lookup
Vector of colors in colorsets}
\usage{
icolors.lookup(color_names = NULL, ...)
}
\arguments{
\item{color_names}{Optional vector of color names; can be partial matches.}

\item{...}{Additional optional color names}
}
\value{
Vector of named RGB colors
}
\description{
icolors.lookup
Vector of colors in colorsets
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link[=icolors.plot]{icolors.plot()}}, \code{\link[=icolors]{icolors()}}
}
\keyword{utilities}
