% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icolors.R
\name{icolors.plot}
\alias{icolors.plot}
\title{icolors.plot
Illustration of icolors
Creates plots of the colors in \code{\link[=icolors]{icolors()}}}
\usage{
icolors.plot(
  set = c("general", "general2", "bg", "bgpng", "cmyk", "CC", "CCalt", "f2", "sex",
    "main", "crayons", "web", "summer21", "pantone", "mario", ""),
  method2order = c("orig", "hsv", "cluster"),
  cex = 0.6,
  mar = rep(1, 4),
  bg = "white",
  fg = "black",
  border = T
)
}
\arguments{
\item{method2order}{method to order colors (\code{'hsv'} or \code{'cluster'})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics:par]{graphics::par()}})}

\item{bg}{Background color}

\item{fg}{Foreground color (for text and box outlines)}

\item{border}{If TRUE, plot a border around each rectangle}
}
\value{
None
}
\description{
icolors.plot
Illustration of icolors
Creates plots of the colors in \code{\link[=icolors]{icolors()}}
}
\examples{
icolors.plot()
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link[=icolors]{icolors()}}
}
