% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icolors.R
\name{icolors}
\alias{icolors}
\title{icolors
Vectors of colors for figures}
\usage{
icolors(
  set = c("general", "general2", "bg", "bgpng", "cmyk", "CC", "CCalt", "f2", "sex",
    "main", "crayons", "web", "summer21", "pantone", "mario", "gothic")
)
}
\arguments{
\item{set}{Character string indicating a set of colors.}
}
\value{
Vector of character strings representing the chosen set of colors, in RGB.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
\examples{
\dontrun{
points(1, 8, pch=21, bg=icolors('main'), cex=4)
}
}
\seealso{
\code{\link[=plot_crayons]{plot_crayons()}}
}
\keyword{utilities}
