% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExtraPackages.R
\name{loadExtraPackages}
\alias{loadExtraPackages}
\title{loadExtraPackages}
\usage{
loadExtraPackages()
}
\description{
\code{loadExtrapackages} loads a set of often used packages for data analysis.
}
\details{
The following packages are loaded:
\itemize{
\item ggplo2
\item tidyverse
\item scales
\item janitor
\item stringr
\item RColorBrewer
\item lubridate
\item rpart
\item rpart.plot
\item vroom
\item sqldf
\item gt
\item patchwork
}

All packages are loaded via \code{library()}, hence if a package can not be loaded the function aborts with an error.
}
