% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByMax.R
\name{normalizeByMax}
\alias{normalizeByMax}
\title{Normalize to the range \link{0, 1}}
\usage{
normalizeByMax(x)
}
\arguments{
\item{x}{a numeric or integer vector}
}
\value{
returns the vector x where each element of x is divided by max(x)
}
\description{
Normalizes a numeric vector to the range \link{0, 1} by dividing each element of x by max(x)
}
\details{
missing values in x are ignored
}
\examples{
normalizeByMax(1:5)
normalizeByMax(c(1:5, NA))
normalizeByMax(letters[1:5])
}
