% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(
  df,
  var1,
  var2,
  perc,
  title = paste("Heatmap - ", var1, "vs", var2),
  legendtitle = "Percentage of\\nObservations",
  col.low = "white",
  col.high = "red"
)
}
\arguments{
\item{df}{todo}

\item{var1}{todo}

\item{var2}{todo}

\item{perc}{todo}

\item{title}{todo}

\item{legendtitle}{todo}

\item{col.low}{todo}

\item{col.high}{todo}
}
\value{
returns a ggplot2 object
}
\description{
Plot Heatmap
}
\examples{
require(ggplot2)
data(diamonds)
df.plot <- diamonds \%>\% group_by(color, cut) \%>\% summarise(nobs=n()) \%>\%
  ungroup() \%>\% mutate(percobs = nobs / sum(nobs))
plotHeatmap(df.plot, "color", "cut", "percobs")
}
