% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapAggr.R
\name{plotHeatmapAggr}
\alias{plotHeatmapAggr}
\title{Plot Heatmap including pre-aggregation of data}
\usage{
plotHeatmapAggr(
  df,
  var1,
  var2,
  col.low = "white",
  col.high = "red",
  title = paste("Heatmap - ", var1, "vs", var2),
  legendtitle = "Percentage of\\nObservations",
  var1.breaks,
  var2.breaks
)
}
\arguments{
\item{df}{x}

\item{var1}{x}

\item{var2}{x}

\item{col.low}{x}

\item{col.high}{x}

\item{title}{x}

\item{legendtitle}{x}

\item{var1.breaks}{x}

\item{var2.breaks}{x}
}
\value{
returns a ggplot2 object
}
\description{
Plot Heatmap including pre-aggregation of data
}
\examples{
require(ggplot2)
data(diamonds)

plotHeatmapAggr(diamonds
                , var1 = "color"
                , var2 = "cut"
                , legendtitle = "Percentage of\nObservations"
)
}
