% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{spam.state}
\alias{spam.state}
\title{spam.state}
\usage{
spam.state(st)
}
\arguments{
\item{state}{Character string indicating a set of colors.}
}
\value{
Printed map of selected state(s) in outline
}
\description{
Creates quick maps of states
}
\examples{
\dontrun{
spam.state(c('california'))
spam.state(c('rhode island', 'new hampshire'))
}

}
