% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{spam.usa}
\alias{spam.usa}
\title{spam.usa}
\usage{
spam.usa()
}
\arguments{
\item{state}{Character string indicating a set of colors.}
}
\value{
Printed map of US in outline
}
\description{
Creates quick maps of U.S.
}
\examples{
\dontrun{
spam.usa()}
}
